requirements.txt
scikit-learn==1.3.0
torch==2.0.0
shap==0.42.1
pandas==2.0.0
numpy==1.24.0
matplotlib==3.7.0
seaborn==0.12.0
psutil==5.9.0
dtreeviz==2.2.0
pytorch-tabnet==4.1.0
imbalanced-learn==0.11.0
scipy==1.11.0
pyyaml==6.0
tqdm==4.65.0
